Fix hover transparency on close button
Overthinking on my part. We can set the opacity to 1 on the titlebar button hover state, without obscuring the icons. When hovering the close button, the background is now fully opaque. This fixes the icon "misalignment" issue for the close button on hover. While technically still "mislaigned", its not noticeable to the end user.

Additionally I've done some testing at various display scales to ensure this is all working correctly. This will probably be the last commit, but I am fairly sure this is a solid CSS solution to the Windows 11 titlebar button issue.